Cleaned up imports, added FileProvider to AndroidManifest.xml to give the camera app permission to add temp files for take picture functionality, split up UiState views into two different composables
