+++ app/src/main/java/com/oracle/visualize/domain/repositories/UserRepository.kt
@@ -1,6 +1,4 @@
 package com.oracle.visualize.domain.repositories
-
-import android.net.Uri
 import com.oracle.visualize.domain.models.ShareUser
 import com.oracle.visualize.domain.models.User
