+++ app/src/main/java/com/oracle/visualize/presentation/components/BasicDialog.kt
@@ -1,0 +1,32 @@
+package com.oracle.visualize.presentation.components
+
+import androidx.compose.material3.AlertDialog
+import androidx.compose.material3.Text
+import androidx.compose.material3.TextButton
+import androidx.compose.runtime.Composable
+
+@Composable
+fun BasicDialog(
+    title: String,
+    message: String,
+    confirm: String,
+    cancel: String,
+    onConfirm: () -> Unit,
+    onDismiss: () -> Unit
+) {
+    AlertDialog(
+        onDismissRequest = onDismiss,
+        title = { Text(title) },
+        text = { Text(message) },
+        confirmButton = {
+            TextButton(onClick = onConfirm) {
+                Text(confirm)
+            }
+        },
+        dismissButton = {
+            TextButton(onClick = onDismiss) {
+                Text(cancel)
+            }
+        }
+    )
+}
