+++ app/src/main/java/com/oracle/visualize/presentation/screens/profileScreen/ProfileView.kt
@@ -58,10 +58,13 @@ import androidx.navigation.NavController
 import coil3.compose.AsyncImage
 import com.oracle.visualize.R
 import com.oracle.visualize.presentation.components.AppDropdownMenu
+import com.oracle.visualize.presentation.components.BasicDialog
 import com.oracle.visualize.presentation.navigation.NavRoutes
 import com.oracle.visualize.presentation.screens.profileScreen.components.ChartThemePicker
 import com.oracle.visualize.presentation.screens.profileScreen.components.ProfileHeader
 import com.oracle.visualize.presentation.screens.profileScreen.components.SettingsCard
+import com.oracle.visualize.presentation.screens.profileScreen.views.EditPfp
+import com.oracle.visualize.presentation.screens.profileScreen.views.EditProfile
 import com.yalantis.ucrop.UCrop
 import java.io.File
 
@@ -193,226 +196,62 @@ fun ProfilePage(
                 }
             }
             is ProfileUiState.Ready -> {
-
-                // Background Image Setup
-                Image(
-                    painter = painterResource(id = R.drawable.profilebgtransparent),
-                    contentDescription = null,
-                    modifier = Modifier
-                        .fillMaxWidth()
-                        .height(180.dp),
-                    alignment = Alignment.TopCenter,
-                    contentScale = ContentScale.FillBounds
+                EditProfile(
+                    appversion = appVersion,
+                    username = state.username,
+                    email = state.eMail,
+                    image = state.image,
+                    chartTheme = state.chartTheme,
+                    onTakePhoto = {
+                        val uri = FileProvider.getUriForFile(
+                            context,
+                            "${context.packageName}.provider",
+                            File.createTempFile("photo_", ".jpg", context.cacheDir)
+                        )
+                        imageUri = uri
+                        if (ContextCompat.checkSelfPermission(context, Manifest.permission.CAMERA)
+                            == PackageManager.PERMISSION_GRANTED) {
+                            cameraLauncher.launch(uri)
+                        } else {
+                            permissionLauncher.launch(Manifest.permission.CAMERA)
+                        }
+                    },
+                    onChoosePhoto = {
+                        galleryLauncher.launch(
+                            PickVisualMediaRequest(ActivityResultContracts.PickVisualMedia.ImageOnly)
+                        )
+                    },
+                    onDeletePhoto = { showDeletePhotoDialog = true },
+                    onPaletteChange = { profileViewModel.setChartTheme(it) },
+                    onLogoutClick = { showLogoutDialog = true }
                 )
-
-                // Page content starts here.
-                Column(
-                    modifier = Modifier
-                        .fillMaxSize()
-                        .padding(horizontal = 16.dp)
-                ) {
-                    Spacer(modifier = Modifier.height(98.dp))
-
-                    var expanded by remember { mutableStateOf(false) }
-
-                    Box {
-                        ProfileHeader(
-                            userName = state.username,
-                            email = state.eMail,
-                            profileImageUrl = state.image,
-                            onEditClick = { expanded = true }
-                        )
-
-                        AppDropdownMenu(
-                            expanded = expanded,
-                            onDismiss = { expanded = false },
-                            items = listOf(
-                                stringResource(R.string.take_photo) to {
-
-                                    // Creates a temporary cache file to store the captured image.
-
-                                    val uri = FileProvider.getUriForFile(
-                                        context,
-                                        "${context.packageName}.provider",
-                                        File.createTempFile("photo_", ".jpg", context.cacheDir)
-                                    )
-
-
-                                    imageUri = uri
-                                    if (ContextCompat.checkSelfPermission(context, Manifest.permission.CAMERA)
-                                        == PackageManager.PERMISSION_GRANTED) {
-                                        cameraLauncher.launch(uri)
-                                    } else {
-                                        permissionLauncher.launch(Manifest.permission.CAMERA)
-                                    }
-
-
-                                },
-                                stringResource(R.string.choose_photo) to {
-                                    galleryLauncher.launch(
-                                        PickVisualMediaRequest(ActivityResultContracts.PickVisualMedia.ImageOnly)
-                                    )
-                                },
-                                stringResource(R.string.delete_photo) to { showDeletePhotoDialog = true }
-                            )
-                        )
-                    }
-
-                    Spacer(modifier = Modifier.height(32.dp))
-
-                    ChartThemePicker(
-                        selectedPalette = state.chartTheme,
-                        onPaletteChange = { profileViewModel.setChartTheme(it) }
-                    )
-
-                    Spacer(modifier = Modifier.height(4.dp))
-
-                    SettingsCard(title = stringResource(R.string.about_title)) {
-                        Text(
-                            text = "${stringResource(R.string.version_declaration)} $appVersion\n${stringResource(R.string.developer)}",
-                            style = MaterialTheme.typography.labelSmall,
-                            color = MaterialTheme.colorScheme.onBackground
-                        )
-                        Text(
-                            text = "${stringResource(R.string.terms_of_service)}\n${stringResource(R.string.licenses)}",
-                            style = MaterialTheme.typography.labelSmall,
-                            color = MaterialTheme.colorScheme.onPrimaryContainer
-                        )
-                    }
-
-                    Spacer(modifier = Modifier.height(64.dp))
-
-                    OutlinedButton(
-                        onClick = { showLogoutDialog = true },
-                        border = BorderStroke(2.dp, MaterialTheme.colorScheme.error),
-                        shape = RoundedCornerShape(50),
-                        colors = ButtonDefaults.outlinedButtonColors(
-                            containerColor = MaterialTheme.colorScheme.secondaryContainer,
-                            contentColor = MaterialTheme.colorScheme.error
-                        ),
-                        modifier = Modifier
-                            .fillMaxWidth()
-                            .padding(horizontal = 48.dp)
-                            .height(64.dp)
-                    ) {
-                        Text(
-                            text = stringResource(R.string.log_out),
-                            style = MaterialTheme.typography.titleMedium
-                        )
-                    }
-                }
             }
             is ProfileUiState.PfpUpload -> {
-
-                Column(
-                    horizontalAlignment = Alignment.CenterHorizontally
+                EditPfp(
+                    pfp = state.pfp,
+                    onBack = { showUnsavedChangesDialog = true },
+                    onEditClick = {
+                        state.pfp?.let { pfp ->
+                            val destUri = Uri.fromFile(File(context.cacheDir, "cropped_pfp.jpg"))
+                            val options = UCrop.Options().apply {
+                                setCircleDimmedLayer(true)
+                                setShowCropGrid(false)
+                                setShowCropFrame(false)
+                            }
+                            val cropIntent = UCrop.of(Uri.parse(pfp), destUri)
+                                .withAspectRatio(1f, 1f)
+                                .withMaxResultSize(512, 512)
+                                .withOptions(options)
+                                .getIntent(context)
+                            uCropLauncher.launch(cropIntent)
+                        }
+                    },
+                    onDeleteClick = { showDeletePhotoDialog = true },
+                    onSaveChanges = {
+                        profileViewModel.updatePfp(state.pfp ?: "")
+                        profileViewModel.setUiState()
+                    }
                 )
-                {
-                    TopAppBar(
-                        title = {
-                            Text(
-                                text = stringResource(R.string.preview),
-                                color = MaterialTheme.colorScheme.onPrimaryContainer
-                            )
-                        },
-                        navigationIcon = {
-                            IconButton(onClick = { showUnsavedChangesDialog = true }) {
-                                Icon(
-                                    imageVector = Icons.AutoMirrored.Filled.ArrowBack,
-                                    contentDescription = stringResource(R.string.back),
-                                    tint = MaterialTheme.colorScheme.onPrimaryContainer
-                                )
-                            }
-                        },
-                        colors = TopAppBarDefaults.topAppBarColors(
-                            containerColor = MaterialTheme.colorScheme.primaryContainer
-                        )
-                    )
-                    AsyncImage(
-                        model = state.pfp,
-                        contentDescription = stringResource(R.string.profile_img_description),
-                        modifier = Modifier
-                            .size(200.dp)
-                            .clip(CircleShape),
-                        contentScale = ContentScale.Crop,
-                        placeholder = painterResource(R.drawable.profile_placeholder),
-                        error = painterResource(R.drawable.profile_placeholder),
-                    )
-
-                    OutlinedButton(
-                        onClick = {
-                            state.pfp?.let { pfp ->
-                                val destUri = Uri.fromFile(File(context.cacheDir, "cropped_pfp.jpg"))
-                                val options = UCrop.Options().apply {
-                                    setCircleDimmedLayer(true)
-                                    setShowCropGrid(false)
-                                    setShowCropFrame(false)
-                                }
-
-                                val cropIntent = UCrop.of(Uri.parse(pfp), destUri)
-                                    .withAspectRatio(1f, 1f)
-                                    .withMaxResultSize(512, 512)
-                                    .withOptions(options)
-                                    .getIntent(context)
-                                uCropLauncher.launch(cropIntent)
-                            }
-                        },
-                        border = BorderStroke(2.dp, MaterialTheme.colorScheme.primary),
-                        shape = RoundedCornerShape(50),
-                        colors = ButtonDefaults.outlinedButtonColors(
-                            contentColor = MaterialTheme.colorScheme.primary
-                        ),
-                        modifier = Modifier
-                            .fillMaxWidth()
-                            .padding(horizontal = 48.dp)
-                            .height(64.dp)
-                    ) {
-                        Text(
-                            text = stringResource(R.string.edit),
-                            style = MaterialTheme.typography.titleMedium
-                        )
-                    }
-
-                    OutlinedButton(
-                        onClick = { showDeletePhotoDialog = true },
-                        border = BorderStroke(2.dp, MaterialTheme.colorScheme.error),
-                        shape = RoundedCornerShape(50),
-                        colors = ButtonDefaults.outlinedButtonColors(
-                            contentColor = MaterialTheme.colorScheme.error
-                        ),
-                        modifier = Modifier
-                            .fillMaxWidth()
-                            .padding(horizontal = 48.dp)
-                            .height(64.dp)
-                    ) {
-                        Text(
-                            text = stringResource(R.string.delete),
-                            style = MaterialTheme.typography.titleMedium
-                        )
-                    }
-
-                    Button(
-                        onClick = {
-                            profileViewModel.updatePfp(state.pfp ?: "")
-                            profileViewModel.setUiState()
-                        },
-                        shape = RoundedCornerShape(50),
-                        colors = ButtonDefaults.outlinedButtonColors(
-                            containerColor = MaterialTheme.colorScheme.primary,
-                            contentColor = MaterialTheme.colorScheme.onPrimary
-                        ),
-                        modifier = Modifier
-                            .fillMaxWidth()
-                            .padding(horizontal = 48.dp)
-                            .height(64.dp)
-                    ) {
-                        Text(
-                            text = stringResource(R.string.save_changes),
-                            style = MaterialTheme.typography.titleMedium
-                        )
-                    }
-                }
-
             }
 
             else -> {}
@@ -420,28 +259,3 @@ fun ProfilePage(
     }
 }
 
-@Composable
-fun BasicDialog(
-    title: String,
-    message: String,
-    confirm: String,
-    cancel: String,
-    onConfirm: () -> Unit,
-    onDismiss: () -> Unit
-) {
-    AlertDialog(
-        onDismissRequest = onDismiss,
-        title = { Text(title) },
-        text = { Text(message) },
-        confirmButton = {
-            TextButton(onClick = onConfirm) {
-                Text(confirm)
-            }
-        },
-        dismissButton = {
-            TextButton(onClick = onDismiss) {
-                Text(cancel)
-            }
-        }
-    )
-}
