+++ app/src/main/java/com/oracle/visualize/presentation/screens/profileScreen/ProfileViewModel.kt
@@ -1,28 +1,17 @@
 package com.oracle.visualize.presentation.screens.profileScreen
 
-import android.net.Uri
 import android.util.Log
-import androidx.compose.runtime.mutableStateOf
 import androidx.lifecycle.ViewModel
-import com.oracle.visualize.R
-import com.oracle.visualize.ui.theme.ChartPalette
 import dagger.hilt.android.lifecycle.HiltViewModel
 import jakarta.inject.Inject
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.setValue
 import androidx.lifecycle.viewModelScope
-import com.google.firebase.auth.FirebaseAuth
 import com.oracle.visualize.domain.exceptions.AppError
-import com.oracle.visualize.domain.models.User
-import com.oracle.visualize.domain.repositories.UserRepository
 import com.oracle.visualize.domain.usecases.DeleteProfilePictureUseCase
 import com.oracle.visualize.domain.usecases.GetCurrentUserUseCase
 import com.oracle.visualize.domain.usecases.GetUserByIDUseCase
-import com.oracle.visualize.domain.usecases.LoginUseCase
 import com.oracle.visualize.domain.usecases.LogoutUseCase
 import com.oracle.visualize.domain.usecases.SetChartThemeUseCase
 import com.oracle.visualize.domain.usecases.UpdatePfpUseCase
-import com.oracle.visualize.presentation.screens.createChartScreen.CreateChartUiState
 import kotlinx.coroutines.Job
 import kotlinx.coroutines.delay
 import kotlinx.coroutines.flow.MutableStateFlow
