+++ app/src/main/java/com/oracle/visualize/presentation/screens/profileScreen/components/ProfileHeader.kt
@@ -1,9 +1,6 @@
 package com.oracle.visualize.presentation.screens.profileScreen.components
 
-import android.util.Log
 import androidx.compose.foundation.BorderStroke
-import androidx.compose.foundation.Image
-
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Spacer
@@ -11,7 +8,6 @@ import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.offset
 import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.shape.CircleShape
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.outlined.Edit
 import androidx.compose.material3.Icon
@@ -22,13 +18,9 @@ import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clip
-import androidx.compose.ui.layout.ContentScale
-import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
 import com.oracle.visualize.R
-import coil3.compose.AsyncImage
 import com.oracle.visualize.presentation.components.Avatar
 
 @Composable
@@ -36,7 +28,7 @@ fun ProfileHeader(
     modifier: Modifier = Modifier,
     userName: String,
     email: String,
-    profileImageUrl: Any,
+    profileImageUrl: String,
     onEditClick: () -> Unit
 ) {
     Column(
