+++ app/src/main/java/com/oracle/visualize/presentation/screens/profileScreen/views/EditPfp.kt
@@ -1,0 +1,147 @@
+package com.oracle.visualize.presentation.screens.profileScreen.views
+
+import androidx.compose.foundation.BorderStroke
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.shape.CircleShape
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.automirrored.filled.ArrowBack
+import androidx.compose.material3.Button
+import androidx.compose.material3.ButtonDefaults
+import androidx.compose.material3.ExperimentalMaterial3Api
+import androidx.compose.material3.Icon
+import androidx.compose.material3.IconButton
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.OutlinedButton
+import androidx.compose.material3.Text
+import androidx.compose.material3.TopAppBar
+import androidx.compose.material3.TopAppBarDefaults
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
+import androidx.compose.ui.layout.ContentScale
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.unit.dp
+import coil3.compose.AsyncImage
+import com.oracle.visualize.R
+
+@OptIn(ExperimentalMaterial3Api::class)
+@Composable
+fun EditPfp(
+
+    pfp: String? = null,
+    onBack: () -> Unit,
+    onEditClick: () -> Unit,
+    onDeleteClick: () -> Unit,
+    onSaveChanges: () -> Unit
+
+)
+{
+    Column(
+        horizontalAlignment = Alignment.CenterHorizontally
+    )
+    {
+        TopAppBar(
+            title = {
+                Text(
+                    text = stringResource(R.string.preview),
+                    color = MaterialTheme.colorScheme.onPrimaryContainer
+                )
+            },
+            navigationIcon = {
+                IconButton(onClick = { onBack() }) {
+                    Icon(
+                        imageVector = Icons.AutoMirrored.Filled.ArrowBack,
+                        contentDescription = stringResource(R.string.back),
+                        tint = MaterialTheme.colorScheme.onPrimaryContainer
+                    )
+                }
+            },
+            colors = TopAppBarDefaults.topAppBarColors(
+                containerColor = MaterialTheme.colorScheme.primaryContainer
+            )
+        )
+
+        Spacer(modifier = Modifier.height(64.dp))
+
+        AsyncImage(
+            model = pfp,
+            contentDescription = stringResource(R.string.profile_img_description),
+            modifier = Modifier
+                .size(280.dp)
+                .clip(CircleShape),
+            contentScale = ContentScale.Crop,
+            placeholder = painterResource(R.drawable.profile_placeholder),
+            error = painterResource(R.drawable.profile_placeholder),
+        )
+
+        Spacer(modifier = Modifier.height(64.dp))
+
+        Column(
+            modifier = Modifier.fillMaxWidth(),
+            verticalArrangement = Arrangement.spacedBy(12.dp)
+        ) {
+            OutlinedButton(
+                onClick = { onEditClick() },
+                border = BorderStroke(2.dp, MaterialTheme.colorScheme.primary),
+                shape = RoundedCornerShape(50),
+                colors = ButtonDefaults.outlinedButtonColors(
+                    contentColor = MaterialTheme.colorScheme.primary
+                ),
+                modifier = Modifier
+                    .fillMaxWidth()
+                    .padding(horizontal = 40.dp)
+                    .height(64.dp)
+            ) {
+                Text(
+                    text = stringResource(R.string.edit),
+                    style = MaterialTheme.typography.titleMedium
+                )
+            }
+
+            OutlinedButton(
+                onClick = { onDeleteClick() },
+                border = BorderStroke(2.dp, MaterialTheme.colorScheme.error),
+                shape = RoundedCornerShape(50),
+                colors = ButtonDefaults.outlinedButtonColors(
+                    contentColor = MaterialTheme.colorScheme.error
+                ),
+                modifier = Modifier
+                    .fillMaxWidth()
+                    .padding(horizontal = 40.dp)
+                    .height(64.dp)
+            ) {
+                Text(
+                    text = stringResource(R.string.delete),
+                    style = MaterialTheme.typography.titleMedium
+                )
+            }
+
+            Button(
+                onClick = { onSaveChanges() },
+                shape = RoundedCornerShape(50),
+                colors = ButtonDefaults.outlinedButtonColors(
+                    containerColor = MaterialTheme.colorScheme.primary,
+                    contentColor = MaterialTheme.colorScheme.onPrimary
+                ),
+                modifier = Modifier
+                    .fillMaxWidth()
+                    .padding(horizontal = 40.dp)
+                    .height(64.dp)
+            ) {
+                Text(
+                    text = stringResource(R.string.save_changes),
+                    style = MaterialTheme.typography.titleMedium
+                )
+            }
+        }
+    }
+}
